requirements.txt
pyyaml
pandas
requests
python-dotenv
